fix:로딩 문제 해결
