+++ package.json
@@ -6,6 +6,8 @@
     "@emotion/react": "^11.13.3",
     "@emotion/styled": "^11.13.0",
     "@fontsource/roboto": "^5.0.15",
+    "@fullcalendar/daygrid": "^6.1.15",
+    "@fullcalendar/react": "^6.1.15",
     "@mui/icons-material": "^6.0.2",
     "@mui/material": "^6.1.1",
     "@mui/styled-engine-sc": "^6.0.2",
@@ -17,6 +19,7 @@
     "jwt-decode": "^4.0.0",
     "mui-tel-input": "^6.0.1",
     "react": "^18.3.1",
+    "react-big-calendar": "^1.14.1",
     "react-dom": "^18.3.1",
     "react-grid-layout": "^1.4.4",
     "react-router-dom": "^6.26.2",
