--- src/pages/calendar/Calendar.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Calendar() {
-  return <div>캘린더</div>;
-}
